+++ integrations/azure-devops/CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.9.11 (2026-06-05)
+
+
+### Features
+
+- Added `excludeTagFilter` selector to the `project` kind.
+- Added `excludeTagFilter` selector to the `work items` kind.
+
+
 ## 0.9.10 (2026-06-03)